Geben was going into an infinite loop calling property_get on a hash from the "geben-display-context" window.  This happened any time the hash had a slash character in the key, like in this code:
  my $testMap = {
    '/doTest/*'    => { 'x' => 1 },
  };

Inspecting the '/doTest/*' key for $testMap would cause geben to
repeatedly call property_get like this:

  property_get -i 38888 -d 0 -c 0 -n $testMap->{"/doTest/*"} -p 8866

...with an ever-increasing value for the -i parameter (incrementing once
per call to property_get).  Geben would continue to call property_get
until you forced it to stop (by exiting emacs, or typing M-x geben-stop
in the "GEBEN... process' window).

Note that this occured for any key value that contained a slash--it
did not require the asterisk to be there also.  I am including that only
because it is relevant later in my explanation.

I used wireshark to capture the traffic on port 9000, and compared
the requests from geben to those from Komodo (the ActiveState IDE).

One of the things I noticed was that Komodo was wrapping the -n
parameter in quotes, and escaping the double quotes (") and the
asterisks (*) -- like this:

  property_get -i 38888 -d 0 -c 0 -n "$testMap->{\"/doTest/\*\"}" -p 8866

Also, the response that geben got back in the property_get call had a
name like this:

  name="$testMap-&gt;{"

... where the response Komodo got back looked like this:

  name="$testMap-&gt;{&#34;/doTest/\*&#34;}"

Also, the "type" parameter for geben was blank:

  type=""

... where in the response Komodo got it was not:

  type="HASH"

And finally, the Komodo response had the data for the child property
with this name:

  name="-&gt;{x}"

Geben's response did not have this.

It appeared to me that geben was not getting back the information it
should be.  So after some work on my part, I figured out how to make
geben make the request in the double-quoted-and-escaped format.

This was accomplished by patching the geben-dbgp-command-property-get function.

I have so far done only very rudimentary testing on this patch, but
it seems to have fixed the problem.